Cleaned up and made revisions to doc
The letter "a" should not go before a vowel.
"An" goes before the word that starts with a vowel.
 The word in this case was "OpenStack".
I also added the plural version for IP Addresses.
Since we are referring to multiple IP Addresses,
we need to make sure we add "addresses" after IP.
also removed additional white space before colon

Change-Id: Ibd20045ffb83f1527007b3b4cef4763b305f724f