Begin the installation process
This has minimal hardware requirements (slightly modified from the
old ha-guide) and the first procedural steps for setting up ones
HA environment.  The information is scant and mostly reflects what
is in the old guide.  It should contain all steps to be taken before
one begins to configure the HA controller cluster.

Change-Id: If28d64ba9e6ae3aa13be526f543a26164aea5e8e