Convert Cinder HA doc to RST
This converts the existing documentation about implementing
high availability for Cinder to RST.  It includes some minor
editorial changes but no content enhancements.

This content needs to be updated and expanded after the conversion
is completed.

Also, disable D000 doc8 check since it does not handle the lineos
directive. This is the same fix that was done for openstack-manuals.

Change-Id: I87e18696d5e18833361cfad0c50d20bff616d74f