Move Configure the VIP to different location in docs
In the current draft for the updated HA Guide, the
"Configure the VIP" subsection is included in the
"Installing high availability packages" section.

Given the previous step is just installing the
operating system, it is moved under the
"Configuring the controller for high availability section",
which describes how to install the Pacemaker/Corosync packages.
This is a pre-requisite for the VIP configuration.

Change-Id: Id5ef39dd59ae322aed26133ad873e4642c5195e4
Closes-Bug: #1484477