Fix node arch and install Guide links in HA Guide Hardware setup
The 'neutron' and 'nova-network' links point to the Kilo version of the
Install Guide, as the hardware setup is based around that 3-node
architecture.
The Liberty Install Guide now uses a 2-node architecture. The HA Guide
should be updated to reflect this 2-node architecture, and these two
links should be updated to the Liberty Install Guide.

Change-Id: I36a0fd0605ee9c8837c1f66318550e82c3f74fa2
Closes-bug: #1506374