Fixed tox exec warnings with obsolete deletion
Current tox execution generated total three warnings.
Two warnings are from checkdeletions and checklinks,
both of which are obsolete now.
The other warning is from checklang because bash command
is in /bin, not in testenv:checklang environment.

So this patch removes some tox obsolete sections:
checkdeletions, checklinks and checksyntax as well,
and adds bash to whitelist_externals in checklang.

Change-Id: If514445272ef9fb5e0f182ba8407e195adaa0b62
Closes-Bug: #1518413